# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API Sistem Manajemen Parkir.",
        "title": "Parking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
      "/api/v1/operations/manual-corrections": {
        "get": {
          "description": "Retrieve all manual corrections with pagination support",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Operations"
          ],
          "summary": "Get all manual corrections",
          "parameters": [
            {
              "type": "integer",
              "default": 1,
              "description": "Page number",
              "name": "page",
              "in": "query"
            },
            {
              "type": "integer",
              "default": 10,
              "description": "Items per page",
              "name": "limit",
              "in": "query"
            },
            {
              "type": "string",
              "description": "Search term",
              "name": "search",
              "in": "query"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.PageResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "post": {
          "description": "Create a new manual correction record",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Operations"
          ],
          "summary": "Create a new manual correction",
          "parameters": [
            {
              "description": "Manual correction data",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/requests.CreateManualCorrectionRequest"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
      "/api/v1/operations/manual-corrections/{id}": {
        "get": {
          "description": "Retrieve a specific manual correction by its ID",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Operations"
          ],
          "summary": "Get manual correction by ID",
          "parameters": [
            {
              "type": "string",
              "description": "Manual correction ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "delete": {
          "description": "Delete a specific manual correction by its ID",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Operations"
          ],
          "summary": "Delete a manual correction",
          "parameters": [
            {
              "type": "string",
              "description": "Manual correction ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
...
      "/financials/lost-ticket-fees": {
        "get": {
          "description": "Get list of lost ticket fees with pagination",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Financials"
          ],
          "summary": "Get All Lost Ticket Fees",
          "parameters": [
            {
              "type": "integer",
              "default": 1,
              "description": "Page number",
              "name": "page",
              "in": "query"
            },
            {
              "type": "integer",
              "default": 10,
              "description": "Items per page",
              "name": "limit",
              "in": "query"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.PageResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "post": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Financials"
          ],
          "summary": "Create New Lost Ticket Fee",
          "parameters": [
            {
              "description": "Lost Ticket Fee Request",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/requests.CreateLostTicketFeeRequest"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
      "/financials/lost-ticket-fees/{id}": {
        "get": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Financials"
          ],
          "summary": "Get Lost Ticket Fee By ID",
          "parameters": [
            {
              "type": "string",
              "description": "Lost Ticket Fee Hash ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "delete": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Financials"
          ],
          "summary": "Delete Lost Ticket Fee",
          "parameters": [
            {
              "type": "string",
              "description": "Lost Ticket Fee Hash ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "requests.CreateLostTicketFeeRequest": {
        "type": "object",
        "required": [
          "lost_ticket_fee",
          "original_fee",
          "processed_by",
          "transaction_id"
        ],
        "properties": {
          "lost_ticket_fee": {
            "type": "number",
            "minimum": 0
          },
          "original_fee": {
            "type": "number",
            "minimum": 0
          },
          "processed_by": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "maxLength": 255
          },
          "transaction_id": {
            "type": "string"
          }
        }
      },
      "requests.CreateManualCorrectionRequest": {
        "type": "object",
        "required": [
          "corrected_by",
          "field_corrected",
          "new_value",
          "old_value",
          "reason",
          "transaction_id"
        ],
        "properties": {
          "corrected_by": {
            "type": "string"
          },
          "field_corrected": {
            "type": "string",
            "maxLength": 100
          },
          "new_value": {
            "type": "string",
            "maxLength": 255
          },
          "old_value": {
            "type": "string",
            "maxLength": 255
          },
          "reason": {
            "type": "string",
            "maxLength": 500
          },
          "transaction_id": {
            "type": "string"
                }
            }
        },
...
        "responses.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "responses.Meta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "current_url": {
                    "description": "Navigasi URL",
                    "type": "string"
                },
                "filters": {
                    "type": "object",
                    "additionalProperties": true
                },
                "first_page": {
                    "description": "Navigasi Angka",
                    "type": "integer"
                },
                "first_url": {
                    "type": "string"
                },
                "last_page": {
                    "type": "integer"
                },
                "last_url": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "next_page": {
                    "type": "integer"
                },
                "next_url": {
                    "type": "string"
                },
                "prev_page": {
                    "description": "Pakai pointer biar bisa null kalau gak ada",
                    "type": "integer"
                },
                "prev_url": {
                    "type": "string"
                },
                "search": {
                    "type": "string"
                },
                "sort": {
                    "description": "Search \u0026 Filter",
                    "type": "string"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "responses.PageResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/responses.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}